% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNV_Types.R
\name{SNV_Types}
\alias{SNV_Types}
\title{SNV_Types Function}
\usage{
SNV_Types(vcf, context_length, ref_genome)
}
\arguments{
\item{vcf}{The loaded VCF (Variant Call Format) file containing the mutations.}

\item{context_length}{The length of the context (number of bases) to consider
around (upstream and downstream) each mutation. It must be an odd positive number.}

\item{ref_genome}{The corresponding reference genome (e.g., human genome hg38).}
}
\value{
A vector containing each SNV mutation in the VCF file.
}
\description{
This function extracts SNV (Single Nucleotide Variant) mutations from a VCF
file and generates mutation strings with reference and alternate alleles.
It also considers a context length to extract the upstream and downstream bases flanking the mutation.
}
\examples{
f <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(f, "hg38")[1:100]
ref_genome <- Hsapiens
SNV_Types(vcf, context_length = 3, ref_genome)   

}
